// src/app/model/contract.ts
export class Contract {
  id?: number;
  reference?: string;
  startDate?: Date;
  endDate?: Date;
  contractType?: string;
  risingAmount?: number;
  createdBy?: any;
  updatedBy?: any[];
  equipementsID?: any[];
  clientID?: number;
  clientFirstName?: string;
  clientLastName?: string;
  lastUpdate?: any;
}